Split responsibilities of `StateComputer` into more distinct "services".
