+++ core-rust/state-manager/src/protocol/protocol_updates/protocol_updaters.rs
@@ -6,27 +6,18 @@ use std::sync::Arc;
 use crate::engine_prelude::*;
 
 use crate::protocol::*;
-use crate::transaction::{ExecutionConfigurator, LedgerTransactionValidator};
 
 pub trait ProtocolUpdater {
     /// Executes these state updates associated with the given protocol version
     /// that haven't yet been applied
     /// (hence "remaining", e.g. if node is restarted mid-protocol update).
-    fn execute_remaining_state_updates(
-        &self,
-        network: &NetworkDefinition,
-        database: Arc<DbLock<ActualStateManagerDatabase>>,
-    );
+    fn execute_remaining_state_updates(&self, database: Arc<DbLock<ActualStateManagerDatabase>>);
 }
 
 pub struct NoOpProtocolUpdater;
 
 impl ProtocolUpdater for NoOpProtocolUpdater {
-    fn execute_remaining_state_updates(
-        &self,
-        _network: &NetworkDefinition,
-        _database: Arc<DbLock<ActualStateManagerDatabase>>,
-    ) {
+    fn execute_remaining_state_updates(&self, _database: Arc<DbLock<ActualStateManagerDatabase>>) {
         // no-op
     }
 }
@@ -46,20 +37,11 @@ impl<G: UpdateBatchGenerator> BatchedUpdater<G> {
 }
 
 impl<G: UpdateBatchGenerator> ProtocolUpdater for BatchedUpdater<G> {
-    fn execute_remaining_state_updates(
-        &self,
-        network: &NetworkDefinition,
-        database: Arc<DbLock<ActualStateManagerDatabase>>,
-    ) {
+    fn execute_remaining_state_updates(&self, database: Arc<DbLock<ActualStateManagerDatabase>>) {
         let database = database.lock();
         let mut txn_committer = ProtocolUpdateTransactionCommitter::new(
             self.new_protocol_version.clone(),
             database.deref(),
-            // The costing and logging parameters (of the Engine) are not really used for flash
-            // transactions; let's still pass sane values.
-            // TODO(when we need non-flash transactions): pass the actually configured flags here.
-            ExecutionConfigurator::new(network, true, false),
-            LedgerTransactionValidator::default_from_network(network),
         );
 
         while let Some(next_batch_idx) = txn_committer.next_committable_batch_idx() {
@@ -73,6 +55,20 @@ impl<G: UpdateBatchGenerator> ProtocolUpdater for BatchedUpdater<G> {
                 None => break,
             }
         }
+
+        // if let Some(next_protocol_version) = series_executor.next_protocol_version() {
+        //     self.
+        //     let mut system_commit_request_factory = SystemCommitRequestFactory {
+        //         epoch: series_executor.epoch(),
+        //         timestamp: proposer_timestamp_ms,
+        //         state_version: series_executor.latest_state_version(),
+        //         proof_origin: LedgerProofOrigin::ProtocolUpdate { genesis_opaque_hash },
+        //     };
+        //     let protocol_update_scenarios = self.scenarios_execution_config
+        //         .to_run_after_protocol_update(&next_protocol_version);
+        //     self.execute_scenarios(&mut system_commit_request_factory, protocol_update_scenarios);
+        // }
+        // TODO(wip): scenarios
     }
 }
