+++ core-rust/state-manager/src/staging/cache.rs
@@ -66,6 +66,8 @@
 
 use super::stage_tree::{Accumulator, Delta, DerivedStageKey, StageKey, StageTree};
 use super::ReadableStore;
+use std::ops::{Deref, DerefMut};
+use std::sync::Arc;
 
 use crate::accumulator_tree::storage::{ReadableAccuTreeStore, TreeSlice};
 use crate::engine_prelude::*;
@@ -76,14 +78,76 @@ use crate::staging::overlays::{
 use crate::staging::{
     AccuTreeDiff, HashStructuresDiff, HashUpdateContext, ProcessedTransactionReceipt, StateTreeDiff,
 };
-use crate::transaction::{LedgerTransactionHash, TransactionLogic};
-use crate::{EpochTransactionIdentifiers, ReceiptTreeHash, StateVersion, TransactionTreeHash};
+use crate::transaction::{
+    HasLedgerTransactionHash, LedgerTransactionHash, PreparedLedgerTransaction, TransactionLogic,
+};
+use crate::{
+    ActualStateManagerDatabase, EpochTransactionIdentifiers, LedgerHashes, ReceiptTreeHash,
+    StateVersion, TransactionTreeHash,
+};
 use im::hashmap::HashMap as ImmutableHashMap;
 use itertools::Itertools;
 
 use crate::store::traits::{SubstateNodeAncestryRecord, SubstateNodeAncestryStore};
-use crate::traits::ConfigurableDatabase;
+use crate::traits::{ConfigurableDatabase, QueryableProofStore};
+use node_common::locks::{DbLock, LockFactory, Mutex};
 use slotmap::SecondaryMap;
+
+pub struct ExecutionCacheManager {
+    database: Arc<DbLock<ActualStateManagerDatabase>>,
+    execution_cache: Mutex<ExecutionCache>,
+}
+
+impl ExecutionCacheManager {
+    pub fn new(
+        database: Arc<DbLock<ActualStateManagerDatabase>>,
+        lock_factory: &LockFactory,
+    ) -> Self {
+        let execution_cache = Self::create_clean_execution_cache(database.lock().deref());
+        Self {
+            database,
+            execution_cache: lock_factory
+                .named("execution_cache")
+                .new_mutex(execution_cache),
+        }
+    }
+
+    pub fn clear(&self) {
+        *self.execution_cache.lock() =
+            Self::create_clean_execution_cache(self.database.lock().deref());
+    }
+
+    pub fn access_exclusively(&self) -> impl DerefMut<Target = ExecutionCache> + '_ {
+        self.execution_cache.lock()
+    }
+
+    pub fn find_transaction_root(
+        &self,
+        parent_transaction_root: &TransactionTreeHash,
+        transactions: &[PreparedLedgerTransaction],
+    ) -> Option<TransactionTreeHash> {
+        let execution_cache = self.execution_cache.lock();
+        let mut transaction_root = parent_transaction_root;
+        for transaction in transactions {
+            transaction_root = match execution_cache.get_cached_transaction_root(
+                transaction_root,
+                &transaction.ledger_transaction_hash(),
+            ) {
+                Some(cached) => cached,
+                None => return None,
+            }
+        }
+        Some(*transaction_root)
+    }
+
+    fn create_clean_execution_cache<S: QueryableProofStore>(database: &S) -> ExecutionCache {
+        let current_transaction_root = database
+            .get_latest_proof()
+            .map(|proof| proof.ledger_header.hashes.transaction_root)
+            .unwrap_or_else(|| LedgerHashes::pre_genesis().transaction_root);
+        ExecutionCache::new(current_transaction_root)
+    }
+}
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd)]
 struct TransactionPlacement {
