+++ core-rust/state-manager/src/transaction/mod.rs
@@ -64,6 +64,7 @@
 
 mod executable_logic;
 mod ledger_transaction;
+mod preparation;
 mod preview;
 mod round_update_transaction;
 mod series_execution;
@@ -71,6 +72,7 @@ mod validation;
 
 pub use executable_logic::*;
 pub use ledger_transaction::*;
+pub use preparation::*;
 pub use preview::*;
 pub use round_update_transaction::*;
 pub use series_execution::*;
