+++ core-rust/state-manager/src/transaction/validation.rs
@@ -129,6 +129,7 @@ impl LedgerTransactionValidator {
     }
 
     pub fn validate_flash(
+        // TODO(wip): consider moving out of here or add comment why validator needed
         &self,
         prepared: PreparedLedgerTransaction,
     ) -> ValidatedLedgerTransaction {
